Move some tests from db-adapter test code to verstion-store test code
