+++ versioned/persist/tests/src/main/java/org/projectnessie/versioned/persist/tests/AbstractDatabaseAdapterVersionStoreTest.java
@@ -43,22 +43,6 @@ public abstract class AbstractDatabaseAdapterVersionStoreTest extends AbstractVe
 
   @Nested
   @SuppressWarnings("ClassCanBeStatic")
-  public class SingleBranch extends AbstractSingleBranch {
-    public SingleBranch() {
-      super(AbstractDatabaseAdapterVersionStoreTest.store);
-    }
-  }
-
-  @Nested
-  @SuppressWarnings("ClassCanBeStatic")
-  public class DuplicateTable extends AbstractDuplicateTable {
-    public DuplicateTable() {
-      super(AbstractDatabaseAdapterVersionStoreTest.store);
-    }
-  }
-
-  @Nested
-  @SuppressWarnings("ClassCanBeStatic")
   public class Tracing extends AbstractTracing {
     public Tracing() {
       super(AbstractDatabaseAdapterVersionStoreTest.store);
